+++ app/controllers/application_controller.rb
@@ -14,6 +14,17 @@ class ApplicationController < ActionController::Base
 		redirect_to '/*path'
 	end
 
+  def after_sign_in_path_for(resource)
+    if current_user.plan.name == "mentor"
+      mentor_root_path
+    elsif current_user.plan.name == "mentee"
+      mentee_root_path
+    end
+  end
+  def after_sign_out_path_for(resource)
+    root_path
+  end
+
   private
   def choose_layout
     return "application" if current_user.nil?
