+++ app/views/internships/_form.html.erb
@@ -1,6 +1,7 @@
 <%= simple_form_for @internship do |f| %>
   <%= f.input :title %>
-  <%= f.input :description %>
+  <%= f.input :description, as: :text %>
   <%= f.input :application_link, label: 'Link to application page' %>
+  <%= f.input :deadline %>
   <%= f.button :submit, 'Post position' %>
 <% end %>
