+++ app/views/internships/_internship.html.erb
@@ -1,0 +1,5 @@
+<tr>
+  <td><%= internship.title %></td>
+  <td><%= internship.description %></td>
+  <td><%= internship.deadline %></td>
+</tr>
