+++ app/views/internships/_internships.html.erb
@@ -1,0 +1,19 @@
+<% if internships.empty? %>
+  <div class='alert alert-warning text-center'>No internships available in this category</div>
+<% else %>
+  <div class="table-responsive">
+    <table class='table table-striped'>
+      <thead>
+        <tr>
+          <th width="25%">Title</th>
+          <th width="45%">Description</th>
+          <th width="20%">Deadline</th>
+        </tr>
+      </thead>
+
+      <tbody>
+        <%= render partial: 'internship', collection: internships %>
+      </tbody>
+    </table>
+  </div>
+<% end %>
