+++ app/views/internships/show.html.erb
@@ -1,4 +1,2 @@
-<p id="notice"><%= notice %></p>
-
 <%= link_to 'Edit', edit_internship_path(@internship) %> |
 <%= link_to 'Back', internships_path %>
