+++ config/routes.rb
@@ -1,5 +1,5 @@
 Rails.application.routes.draw do
-  resources :internships
+  resources :internships, except: :show
 
   devise_for :admin_users, ActiveAdmin::Devise.config
   ActiveAdmin.routes(self)
